move types for data migration to types package
