--- lambdas/data-migration2/src/types.ts
@@ -1,11 +1,0 @@
-export interface GranuleDynamoSearchParams {
-  collectionId?: string
-  granuleId?: string
-}
-
-export interface MigrationSummary {
-  dynamoRecords: number
-  success: number
-  skipped: number
-  failed: number
-}
